Turned id into an uint64
Id is now made up of two parts first 16 bits holds the creator id
and the rest holds the order counter. So this way we guarantee that no
two orders will have the same order id.

+++ book/orderlevel.go
@@ -11,7 +11,7 @@ type OrderLevel struct {
 }
 
 func (level *OrderLevel) match(order Order) ([]cti.TradedOrder, Order) {
-  var t []cti.TradedOrder
+	var t []cti.TradedOrder
 	for order.Size > 0 {
 		thisOrder, err := level.Orders.Pop()
 		if err != nil {
@@ -33,13 +33,12 @@ func (level *OrderLevel) match(order Order) ([]cti.TradedOrder, Order) {
 		if order.Side == BUY {
 			to, from = from, to
 		}
-    t = append(t, cti.TradedOrder{
-      To: to,
-      From: from,
-      Size: tradeSize,
-      Price: thisOrder.Price,
-    })
-
+		t = append(t, cti.TradedOrder{
+			To:    to,
+			From:  from,
+			Size:  tradeSize,
+			Price: thisOrder.Price,
+		})
 
 		order.Size -= tradeSize
 		thisOrder.Size -= tradeSize
@@ -109,8 +108,8 @@ func (level *OrderLevel) insert(order Order) *OrderLevel {
 	return level
 }
 
-func (level *OrderLevel) remove(Price float32, id int) (Order, bool) {
-  var o Order
+func (level *OrderLevel) remove(Price float32, id uint64) (Order, bool) {
+	var o Order
 	if level == nil {
 		return o, false
 	}
@@ -121,7 +120,7 @@ func (level *OrderLevel) remove(Price float32, id int) (Order, bool) {
 			level = level.GreaterLevel
 		}
 		if level.Price != Price {
-		return o, false
+			return o, false
 		}
 	} else if level.Price > Price {
 		for level.Price != Price &&
@@ -133,9 +132,9 @@ func (level *OrderLevel) remove(Price float32, id int) (Order, bool) {
 			return o, false
 		}
 	}
-  o, present := level.Orders.Remove(id)
-  if present {
-    level.OrderCount -= o.Size
-  }
-  return o, true
+	o, present := level.Orders.Remove(id)
+	if present {
+		level.OrderCount -= o.Size
+	}
+	return o, true
 }
